Refactoring rest of the common module + fixed its docstrings
